chore: change the name of the functionality
