+++ lib/app/screens/sign_up/sign_up_0111.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:partyguam/app/screens/sign_up/widgets/app_bar.dart';
+import 'package:partyguam/app/screens/sign_up/widgets/buttons.dart';
 import 'package:partyguam/app/screens/sign_up/widgets/components.dart';
+import 'package:partyguam/app/screens/sign_up/widgets/text.dart';
 
 class SignUp0111 extends StatelessWidget {
   const SignUp0111({super.key});
@@ -18,16 +20,16 @@ class SignUp0111 extends StatelessWidget {
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              TitleSection(
+              TitleText(
                 mainTitle: '가입을 축하드려요!\n이메일 정보가 맞나요?',
                 subTitle: '나중에 변경할 수 없어요',
               ),
               // TODO: SSO 구현 후 email 값 변경 필요
-              DisabledInputSection(email: 'partguam@gmail.com'),
+              EmailConfirmForm(email: 'partguam@gmail.com'),
               Expanded(
                 child: SizedBox(),
               ),
-              ButtonSection(
+              MainHorizontalButton(
                 content: '네, 맞아요',
                 routeName: '/sign_up/0112',
               ),
