+++ lib/app/screens/sign_up/sign_up_0113.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:partyguam/app/screens/sign_up/widgets/app_bar.dart';
-import 'package:partyguam/app/screens/sign_up/widgets/components.dart';
+import 'package:partyguam/app/screens/sign_up/widgets/buttons.dart';
+import 'package:partyguam/app/screens/sign_up/widgets/text.dart';
 
 class SignUp0113 extends StatelessWidget {
   const SignUp0113({super.key});
@@ -11,11 +12,11 @@ class SignUp0113 extends StatelessWidget {
       appBar: SignUpAppBar(title: '가입하기', pageCount: '3/4'),
       body: Column(
         children: [
-          TitleSection(
+          TitleText(
               mainTitle: '***님의\n생년월일을 알려주세요',
               subTitle: '프로필에서 노출여부를 설정 할 수 있어요'),
           DateSection(),
-          ButtonSection(content: '다음', routeName: 'sign_up/0114')
+          MainHorizontalButton(content: '다음', routeName: 'sign_up/0114')
         ],
       ),
     );
